+++ src/ipm/ipx/model.cc
@@ -108,9 +108,9 @@ void Model::PresolveStartingPoint(const double* x_user,
                                   Vector& z_solver) const {
     const Int m = rows();
     const Int n = cols();
-    assert(x_solver.size() == n+m);
-    assert(y_solver.size() == m);
-    assert(z_solver.size() == n+m);
+    assert((Int)x_solver.size() == n+m);
+    assert((Int)y_solver.size() == m);
+    assert((Int)z_solver.size() == n+m);
 
     Vector x_temp(num_var_);
     Vector slack_temp(num_constr_);
@@ -220,12 +220,12 @@ void Model::PostsolveInteriorSolution(const Vector& x_solver,
                                       double* zl_user, double* zu_user) const {
     const Int m = rows();
     const Int n = cols();
-    assert(x_solver.size() == n+m);
-    assert(xl_solver.size() == n+m);
-    assert(xu_solver.size() == n+m);
-    assert(y_solver.size() == m);
-    assert(zl_solver.size() == n+m);
-    assert(zu_solver.size() == n+m);
+    assert((Int)x_solver.size() == n+m);
+    assert((Int)xl_solver.size() == n+m);
+    assert((Int)xu_solver.size() == n+m);
+    assert((Int)y_solver.size() == m);
+    assert((Int)zl_solver.size() == n+m);
+    assert((Int)zu_solver.size() == n+m);
 
     Vector x_temp(num_var_);
     Vector xl_temp(num_var_);
@@ -264,12 +264,12 @@ void Model::EvaluateInteriorSolution(const Vector& x_solver,
                                      Info* info) const {
     const Int m = rows();
     const Int n = cols();
-    assert(x_solver.size() == n+m);
-    assert(xl_solver.size() == n+m);
-    assert(xu_solver.size() == n+m);
-    assert(y_solver.size() == m);
-    assert(zl_solver.size() == n+m);
-    assert(zu_solver.size() == n+m);
+    assert((Int)x_solver.size() == n+m);
+    assert((Int)xl_solver.size() == n+m);
+    assert((Int)xu_solver.size() == n+m);
+    assert((Int)y_solver.size() == m);
+    assert((Int)zl_solver.size() == n+m);
+    assert((Int)zu_solver.size() == n+m);
 
     // Build solution to scaled user model.
     Vector x(num_var_);
@@ -302,8 +302,8 @@ void Model::EvaluateInteriorSolution(const Vector& x_solver,
     }
     // rb = rhs-slack-A*x
     Vector rb = scaled_rhs_ - slack;
-    assert(scaled_rhs_.size() == num_constr_);
-    assert(rb.size() == num_constr_);
+    assert((Int)scaled_rhs_.size() == num_constr_);
+    assert((Int)rb.size() == num_constr_);
     MultiplyWithScaledMatrix(x, -1.0, rb, 'N');
     // rc = obj-zl+zu-A'y
     Vector rc = scaled_obj_ - zl + zu;
@@ -361,10 +361,10 @@ void Model::PostsolveBasicSolution(const Vector& x_solver,
                                    double* y_user, double* z_user) const {
     const Int m = rows();
     const Int n = cols();
-    assert(x_solver.size() == n+m);
-    assert(y_solver.size() == m);
-    assert(z_solver.size() == n+m);
-    assert(basic_status_solver.size() == n+m);
+    assert((Int)x_solver.size() == n+m);
+    assert((Int)y_solver.size() == m);
+    assert((Int)z_solver.size() == n+m);
+    assert((Int)basic_status_solver.size() == n+m);
 
     Vector x_temp(num_var_);
     Vector slack_temp(num_constr_);
@@ -395,10 +395,10 @@ void Model::EvaluateBasicSolution(const Vector& x_solver,
                                   Info* info) const {
     const Int m = rows();
     const Int n = cols();
-    assert(x_solver.size() == n+m);
-    assert(y_solver.size() == m);
-    assert(z_solver.size() == n+m);
-    assert(basic_status_solver.size() == n+m);
+    assert((Int)x_solver.size() == n+m);
+    assert((Int)y_solver.size() == m);
+    assert((Int)z_solver.size() == n+m);
+    assert((Int)basic_status_solver.size() == n+m);
 
     // Build basic solution to scaled user model.
     Vector x(num_var_);
@@ -455,7 +455,7 @@ void Model::PostsolveBasis(const std::vector<Int>& basic_status_solver,
                            Int* cbasis_user, Int* vbasis_user) const {
     const Int m = rows();
     const Int n = cols();
-    assert(basic_status_solver.size() == n+m);
+    assert((Int)basic_status_solver.size() == n+m);
 
     std::vector<Int> cbasis_temp(num_constr_);
     std::vector<Int> vbasis_temp(num_var_);
@@ -584,13 +584,13 @@ void Model::ScaleModel(const Control& control) {
 
     // Apply scaling to vectors.
     if (colscale_.size() > 0) {
-        assert(colscale_.size() == num_var_);
+        assert((Int)colscale_.size() == num_var_);
         scaled_obj_ *= colscale_;
         scaled_lbuser_ /= colscale_;
         scaled_ubuser_ /= colscale_;
     }
     if (rowscale_.size() > 0) {
-        assert(rowscale_.size() == num_constr_);
+        assert((Int)rowscale_.size() == num_constr_);
         scaled_rhs_ *= rowscale_;
     }
 }
@@ -1051,13 +1051,13 @@ void Model::DualizeBasicSolution(const Vector& x_user,
 
     if (dualized_) {
         assert(num_var_ == m);
-        assert(num_constr_ + boxed_vars_.size() == n);
+        assert(num_constr_ + (Int)boxed_vars_.size() == n);
 
         // Build dual solver variables from primal user variables.
         y_solver = -x_user;
         for (Int i = 0; i < num_constr_; i++)
             z_solver[i] = -slack_user[i];
-        for (Int k = 0; k < boxed_vars_.size(); k++) {
+        for (Int k = 0; k < (Int)boxed_vars_.size(); k++) {
             Int j = boxed_vars_[k];
             z_solver[num_constr_+k] = c(num_constr_+k) + y_solver[j];
         }
@@ -1067,7 +1067,7 @@ void Model::DualizeBasicSolution(const Vector& x_user,
         // Build primal solver variables from dual user variables.
         std::copy_n(std::begin(y_user), num_constr_, std::begin(x_solver));
         std::copy_n(std::begin(z_user), num_var_, std::begin(x_solver) + n);
-        for (Int k = 0; k < boxed_vars_.size(); k++) {
+        for (Int k = 0; k < (Int)boxed_vars_.size(); k++) {
             Int j = boxed_vars_[k];
             if (x_solver[n+j] < 0.0) {
                 // j is a boxed variable and z_user[j] < 0
@@ -1171,7 +1171,7 @@ void Model::DualizeBackInteriorSolution(const Vector& x_solver,
 
     if (dualized_) {
         assert(num_var_ == m);
-        assert(num_constr_ + boxed_vars_.size() == n);
+        assert(num_constr_ + (Int)boxed_vars_.size() == n);
         x_user = -y_solver;
 
         // If the solution from the solver would be exact, we could copy the
@@ -1310,7 +1310,7 @@ void Model::DualizeBackBasicSolution(const Vector& x_solver,
 
     if (dualized_) {
         assert(num_var_ == m);
-        assert(num_constr_ + boxed_vars_.size() == n);
+        assert(num_constr_ + (Int)boxed_vars_.size() == n);
         x_user = -y_solver;
         for (Int i = 0; i < num_constr_; i++)
             slack_user[i] = -z_solver[i];
@@ -1340,7 +1340,7 @@ void Model::DualizeBackBasis(const std::vector<Int>& basic_status_solver,
 
     if (dualized_) {
         assert(num_var_ == m);
-        assert(num_constr_ + boxed_vars_.size() == n);
+        assert(num_constr_ + (Int)boxed_vars_.size() == n);
         for (Int i = 0; i < num_constr_; i++) {
             if (basic_status_solver[i] == IPX_basic)
                 cbasis_user[i] = IPX_nonbasic;
@@ -1402,8 +1402,8 @@ void Model::CorrectScaledBasicSolution(Vector& x, Vector& slack, Vector& y,
 void Model::MultiplyWithScaledMatrix(const Vector& rhs, double alpha,
                                      Vector& lhs, char trans) const {
     if (trans == 't' || trans == 'T') {
-        assert(rhs.size() == num_constr_);
-        assert(lhs.size() == num_var_);
+        assert((Int)rhs.size() == num_constr_);
+        assert((Int)lhs.size() == num_var_);
         if (dualized())
             for (Int i = 0; i < num_constr_; i++)
                 ScatterColumn(AI_, i, alpha*rhs[i], lhs);
@@ -1412,8 +1412,8 @@ void Model::MultiplyWithScaledMatrix(const Vector& rhs, double alpha,
                 lhs[j] += alpha * DotColumn(AI_, j, rhs);
     }
     else {
-        assert(rhs.size() == num_var_);
-        assert(lhs.size() == num_constr_);
+        assert((Int)rhs.size() == num_var_);
+        assert((Int)lhs.size() == num_constr_);
         if (dualized())
             for (Int i = 0; i < num_constr_; i++)
                 lhs[i] += alpha * DotColumn(AI_, i, rhs);
@@ -1429,7 +1429,7 @@ double PrimalInfeasibility(const Model& model, const Vector& x) {
     assert(x.size() == lb.size());
 
     double infeas = 0.0;
-    for (Int j = 0; j < x.size(); j++) {
+    for (size_t j = 0; j < x.size(); j++) {
         infeas = std::max(infeas, lb[j]-x[j]);
         infeas = std::max(infeas, x[j]-ub[j]);
     }
@@ -1444,7 +1444,7 @@ double DualInfeasibility(const Model& model, const Vector& x,
     assert(z.size() == lb.size());
 
     double infeas = 0.0;
-    for (Int j = 0; j < x.size(); j++) {
+    for (size_t j = 0; j < x.size(); j++) {
         if (x[j] > lb[j])
             infeas = std::max(infeas, z[j]);
         if (x[j] < ub[j])
@@ -1456,10 +1456,10 @@ double DualInfeasibility(const Model& model, const Vector& x,
 double PrimalResidual(const Model& model, const Vector& x) {
     const SparseMatrix& AIt = model.AIt();
     const Vector& b = model.b();
-    assert(x.size() == AIt.rows());
+    assert((Int)x.size() == AIt.rows());
 
     double res = 0.0;
-    for (Int i = 0; i < b.size(); i++) {
+    for (Int i = 0; i < (Int)b.size(); i++) {
         double r = b[i] - DotColumn(AIt, i, x);
         res = std::max(res, std::abs(r));
     }
@@ -1469,11 +1469,11 @@ double PrimalResidual(const Model& model, const Vector& x) {
 double DualResidual(const Model& model, const Vector& y, const Vector& z) {
     const SparseMatrix& AI = model.AI();
     const Vector& c = model.c();
-    assert(y.size() == AI.rows());
-    assert(z.size() == AI.cols());
+    assert((Int)y.size() == AI.rows());
+    assert((Int)z.size() == AI.cols());
 
     double res = 0.0;
-    for (Int j = 0; j < c.size(); j++) {
+    for (Int j = 0; j < (Int)c.size(); j++) {
         double r = c[j] - z[j] - DotColumn(AI, j, y);
         res = std::max(res, std::abs(r));
     }
